+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.1.1
 
 * Added tables support.
+* Made a code block scrollable.
+* Changed images span default alignment.
 
 ## 0.1.0